+++ common/src/main/java/io/github/mortuusars/envelope/world/entity/ai/goal/PigeonDeliverMailGoal.java
@@ -16,6 +16,11 @@ public class PigeonDeliverMailGoal extends Goal {
     }
 
     @Override
+    public boolean requiresUpdateEveryTick() {
+        return true;
+    }
+
+    @Override
     public boolean canUse() {
         return pigeon.isDelivering();
     }