+++ common/src/main/java/io/github/mortuusars/envelope/world/mail/entity/VillagerMailEntity.java
@@ -10,8 +10,8 @@ import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.Nullable;
 
 public class VillagerMailEntity extends MailEntity {
-    public VillagerMailEntity(Address.Entity address) {
-        super(address);
+    public VillagerMailEntity(Address.Entity address, int distance) {
+        super(address, distance);
     }
 
     @Override